feature: 추천 게시글 수정 완료
