+++ service/src/apis/HttpClient.tsx
@@ -1,8 +1,5 @@
 import axios from 'axios';
 
-const INITIAL_TOKEN = 'accessToken';
-const getToken = localStorage.getItem(INITIAL_TOKEN);
-
 const baseURL = import.meta.env.VITE_APP_BACK_END_LOCAL;
 
 const HttpClient = axios.create({
@@ -13,6 +10,8 @@ const HttpClient = axios.create({
 });
 HttpClient.interceptors.request.use(
     async config => {
+        const INITIAL_TOKEN = 'accessToken';
+        const getToken = localStorage.getItem(INITIAL_TOKEN);
         const token = getToken;
         if (token !== null) {
             config.headers.Authorization = `Bearer ${token}`;
