+++ service/src/components/aside/Aside.tsx
@@ -2,13 +2,31 @@ import TopPost from './TopPost';
 import Carousel from '../carousel/Carousel';
 import TopKeywords from './TopKeywords';
 import { recommendState } from '../../recoil/atom/recommendState';
-import { useRecoilValue } from 'recoil';
+import { useRecoilValue, useRecoilState } from 'recoil';
 import { isLoggedInState } from '../../recoil/atom/isLoggedInState';
+import { useTokenDecode } from '../../hooks/useTokenDecode';
+import { getAuthStorage } from '../../repository/AuthRepository';
+import { getRecommendTechBlogService } from '../../service/TechBlogService';
+import { useEffect } from 'react';
 
 export default function Aside() {
-    const loggedIn = useRecoilValue(isLoggedInState);
+    const [recommendData, setRecommendData] = useRecoilState(recommendState);
+    const TOKEN_KEY = 'accessToken';
 
-    const recommendData = useRecoilValue(recommendState);
+    const loggedIn = useRecoilValue(isLoggedInState);
+    const token = getAuthStorage(TOKEN_KEY);
+    const memberId = useTokenDecode(token);
+
+    async function getRecommenedDataRender() {
+        if (memberId !== undefined) {
+            const recommendBlogData = await getRecommendTechBlogService(memberId);
+            setRecommendData(recommendBlogData);
+        }
+    }
+    useEffect(() => {
+        getRecommenedDataRender();
+    }, [loggedIn]);
+
     return (
         <div className="flex flex-col justify-around w-full p-4 my-10">
             {loggedIn && (
