+++ service/src/components/layout/Layout.tsx
@@ -2,29 +2,29 @@ import { Outlet } from 'react-router-dom';
 
 import Header from '@monorepo/component/src/stories/header/Header';
 import Aside from '../aside/Aside';
-import { getAuthStorage } from '../../repository/AuthRepository';
-import { useEffect } from 'react';
-import { useTokenDecode } from '../../hooks/useTokenDecode';
-import { getRecommendTechBlogService } from '../../service/TechBlogService';
-import { useSetRecoilState, useRecoilValue } from 'recoil';
-import { recommendState } from '../../recoil/atom/recommendState';
-import { isLoggedInState } from '../../recoil/atom/isLoggedInState';
+// import { getAuthStorage } from '../../repository/AuthRepository';
+// import { useEffect } from 'react';
+// import { useTokenDecode } from '../../hooks/useTokenDecode';
+// import { getRecommendTechBlogService } from '../../service/TechBlogService';
+// import { useSetRecoilState, useRecoilValue } from 'recoil';
+// import { recommendState } from '../../recoil/atom/recommendState';
+// import { isLoggedInState } from '../../recoil/atom/isLoggedInState';
 
 export default function Layout() {
-    const setRecommendData = useSetRecoilState(recommendState);
-    const TOKEN_KEY = 'accessToken';
-    const token = getAuthStorage(TOKEN_KEY);
-    const memberId = useTokenDecode(token);
-    const loggedIn = useRecoilValue(isLoggedInState);
+    // const setRecommendData = useSetRecoilState(recommendState);
+    // const TOKEN_KEY = 'accessToken';
+    // const token = getAuthStorage(TOKEN_KEY);
+    // const memberId = useTokenDecode(token);
+    // const loggedIn = useRecoilValue(isLoggedInState);
 
-    async function getRecommenedDataRender() {
-        const recommendBlogData = await getRecommendTechBlogService(memberId);
-        setRecommendData(recommendBlogData);
-    }
+    // async function getRecommenedDataRender() {
+    //     const recommendBlogData = await getRecommendTechBlogService(memberId);
+    //     setRecommendData(recommendBlogData);
+    // }
 
-    useEffect(() => {
-        getRecommenedDataRender();
-    }, [loggedIn]);
+    // useEffect(() => {
+    //     getRecommenedDataRender();
+    // }, [loggedIn, token]);
 
     return (
         <>
