+++ service/src/service/TechBlogService.ts
@@ -1,5 +1,4 @@
 import HttpClient from '../apis/HttpClient';
-import { getAuthStorage } from '../repository/AuthRepository';
 
 interface ITechBlogCategory {
     page: number;
@@ -39,10 +38,6 @@ export async function getUserTechBlogService({
 
 export async function getRecommendTechBlogService(memberId: string) {
     try {
-        const TOKEN_KEY = 'accessToken';
-        const token = getAuthStorage(TOKEN_KEY);
-        console.log(token);
-
         const res = await HttpClient.get(`/api/v1/recommendation/posts?${memberId}`);
         return res.data;
     } catch (error) {
